Update to hashing package and other stuff

+++ Project/src/main/java/domain/DomainFacade.java
@@ -3,9 +3,8 @@ package domain;
 import Interfaces.*;
 import data.PersistenceFacade;
 import domain.CreditsManagement.CreditsSystem;
-import domain.authentication.AuthenticationHandler;
+import domain.authentication.AuthenticationFacade;
 import domain.notification.ProducerNotification;
-import domain.notification.AdminNotification;
 import domain.searchEngine.SearchEngineHandler;
 import domain.searchEngine.SearchUserHandler;
 import domain.session.CurrentSession;
@@ -33,7 +32,7 @@ public class DomainFacade implements IDomainFacade {
 
     @Override
     public boolean login(IUser user) {
-        return AuthenticationHandler.getLoginInstance().login(user);
+        return AuthenticationFacade.getInstance().login(user);
     }
 
     public void addCredit(IProduction production, IRightsholder rightsholder, List<String> roles) {
@@ -121,7 +120,7 @@ public class DomainFacade implements IDomainFacade {
 
     @Override
     public boolean validateUser(IUser user) {
-        return AuthenticationHandler.getUserInstance().validateUser(user);
+        return AuthenticationFacade.getInstance().validateUser(user);
     }
 
 
@@ -132,12 +131,20 @@ public class DomainFacade implements IDomainFacade {
 
     @Override
     public boolean deleteUser(IUser user) {
-        return PersistenceFacade.getInstance().deleteUser(user);
+        IUser currentUser = getCurrentUser();
+        if (validateUser(currentUser) && user != null && !currentUser.getUsername().equals(user.getUsername())) {
+            return PersistenceFacade.getInstance().deleteUser(user);
+        }
+        return false;
     }
 
     @Override
     public boolean editUser(IUser user) {
-        return PersistenceFacade.getInstance().editUser(user);
+        IUser currentUser = getCurrentUser();
+        if (validateUser(currentUser) && !user.getUsername().equals("") && !user.getPassword().equals("")) {
+            return PersistenceFacade.getInstance().editUser(user);
+        }
+        return false;
     }
 
     @Override
@@ -147,7 +154,13 @@ public class DomainFacade implements IDomainFacade {
 
     @Override
     public boolean addUser(IUser user) {
-        if (PersistenceFacade.getInstance().getUser(user) == null) {
+        IUser currentUser = getCurrentUser();
+        if (validateUser(currentUser) && !user.getUsername().equals("") && !user.getPassword().equals("")) {
+            try {
+                user.setPassword(AuthenticationFacade.getInstance().generateStrongPasswordHash(user.getPassword()));
+            } catch (NoSuchAlgorithmException | InvalidKeySpecException e) {
+                e.printStackTrace();
+            }
             return PersistenceFacade.getInstance().addUser(user);
         }
         return false;
@@ -228,9 +241,10 @@ public class DomainFacade implements IDomainFacade {
         return PersistenceFacade.getInstance().countUnreadProducerNotifications(user);
     }
 
+    @Override
     public String generateStrongPasswordHash(String password) {
         try {
-            return AuthenticationHandler.getInstance().generateStrongPasswordHash(password);
+            return AuthenticationFacade.getInstance().generateStrongPasswordHash(password);
         } catch (NoSuchAlgorithmException | InvalidKeySpecException e) {
             e.printStackTrace();
         }
